refactor(imports): replace deep submodule imports with public API imports (#2446)
Replace 42 deep cross-module imports (from vibe3.X.submodule import Y) with
top-level public API imports (from vibe3.X import Y) across 10 target modules.
All 67 imported symbols are already re-exported in their respective module's
__all__, making this a mechanical find-and-replace operation.

10 imports required merging into existing top-level import lines. 2 lazy-imported
GITHUB_* constants moved to function-body imports using public API path to
preserve lazy loading through __getattr__.

Modules: clients, domain, environment, observability, prompts, analysis, utils,
runtime, server, ui.
