+++ src/vibe3/analysis/inspect_query_service.py
@@ -32,8 +32,7 @@ from . import dag_service as dag_service_module
 
 def build_change_analysis(source_type: str, identifier: str) -> dict[str, object]:
     """Run change analysis pipeline and return impact/dag/score."""
-    from vibe3.clients.git_client import GitClient
-    from vibe3.clients.github_client import GitHubClient
+    from vibe3.clients import GitClient, GitHubClient
 
     log = logger.bind(
         domain="inspect", action="change_analysis", source_type=source_type
