+++ src/vibe3/domain/events/base.py
@@ -4,6 +4,6 @@ Re-exported from models layer to break L3 circular dependencies.
 See vibe3.models.domain_events for the canonical definition.
 """
 
-from vibe3.models.domain_events import DomainEvent
+from vibe3.models import DomainEvent
 
 __all__ = ["DomainEvent"]
