+++ src/vibe3/domain/handlers/issue_state_dispatch.py
@@ -131,12 +131,12 @@ def handle_manager_dispatch_intent(
                 state=target_state,
             )
         else:
-            from vibe3.clients.github_field_constants import GITHUB_DEFAULT_VIEW_FIELDS
+            from vibe3.clients import GITHUB_DEFAULT_VIEW_FIELDS
 
             issue_data = await loop.run_in_executor(
                 None,
                 lambda: ctx.github_client.view_issue(
-                    event.issue_number, fields=list(GITHUB_DEFAULT_VIEW_FIELDS)
+                    event.issue_number, fields=list(GITHUB_DEFAULT_VIEW_FIELDS)  # type: ignore[call-overload]
                 ),
             )
