+++ src/vibe3/domain/publisher.py
@@ -4,7 +4,7 @@ Re-exported from models layer to break L3 circular dependencies.
 See vibe3.models.event_bus for the canonical implementation.
 """
 
-from vibe3.models.event_bus import (
+from vibe3.models import (
     EventHandler,
     EventPublisher,
     get_publisher,
