+++ src/vibe3/runtime/orchestra_instance.py
@@ -4,7 +4,7 @@ Re-exports from vibe3.utils.orchestra_instance.
 Actual implementation moved to utils/ to break services→runtime circular dependency.
 """
 
-from vibe3.utils.orchestra_instance import (
+from vibe3.utils import (
     OrchestraInstanceInfo,
     read_instance_info,
     validate_instance,
