+++ src/vibe3/server/mcp.py
@@ -8,8 +8,7 @@ from loguru import logger
 if TYPE_CHECKING:
     from mcp.server.fastmcp import FastMCP
 
-    from vibe3.services import OrchestraStatusService
-    from vibe3.services.orchestra_status_service import OrchestraSnapshot
+    from vibe3.services import OrchestraSnapshot, OrchestraStatusService
 
 
 def _serialize_snapshot(snapshot: "OrchestraSnapshot") -> dict:
