+++ src/vibe3/ui/flow_ui.py
@@ -2,8 +2,15 @@
 
 from typing import Any
 
-from vibe3.models.flow import FlowStatusResponse
-from vibe3.services import ref_to_handoff_cmd
+from vibe3.models import FlowStatusResponse
+from vibe3.services import (
+    FlowCategory,
+    FlowState,
+    SpecRefService,
+    classify_flow,
+    get_flow_state,
+    ref_to_handoff_cmd,
+)
 from vibe3.ui.console_impl import console
 from vibe3.ui.flow_ui_primitives import display_actor, kv, resolve_ref_path, status_text
 from vibe3.ui.flow_ui_timeline import render_flow_timeline  # noqa: F401
@@ -157,8 +164,6 @@ def render_flow_status(
             f"  [dim]verdict:[/] [{verdict_color}]{v.verdict}[/]" f"  [dim]{v.actor}[/]"
         )
 
-    from vibe3.services.spec_ref_service import SpecRefService
-
     spec_service = SpecRefService()
     spec_display = spec_service.get_spec_display(
         status.spec_ref, status.task_issue_number
@@ -240,13 +245,6 @@ def render_flows_status_dashboard(
     - Blocked: Flows with blocked_by
     - Done/Aborted/Stale: Completed flows
     """
-    from vibe3.services.flow_classifier import (
-        FlowCategory,
-        FlowState,
-        classify_flow,
-        get_flow_state,
-    )
-
     pr_map = pr_map or {}
     worktree_map = worktree_map or {}
